+++ pom.xml
@@ -23,13 +23,25 @@
     <properties>
         <maven.compiler.source>8</maven.compiler.source>
         <maven.compiler.target>8</maven.compiler.target>
-        <rocketmq-streams.version>1.1.0-preview-SNAPSHOT</rocketmq-streams.version>
+        <rocketmq-streams.version>1.1.0</rocketmq-streams.version>
         <antlr.version>4.9.2</antlr.version>
         <logback.version>1.2.10</logback.version>
     </properties>
 
     <dependencyManagement>
         <dependencies>
+            <dependency>
+                <groupId>org.apache.rocketmq</groupId>
+                <artifactId>rocketmq-streams</artifactId>
+                <exclusions>
+                    <exclusion>
+                        <groupId>ch.qos.logback</groupId>
+                        <artifactId>logback-core</artifactId>
+                    </exclusion>
+                </exclusions>
+                <version>${rocketmq-streams.version}</version>
+            </dependency>
+
             <dependency>
                 <groupId>com.alibaba</groupId>
                 <artifactId>rsqldb-common</artifactId>
