+++ rsqldb-rest/pom.xml
@@ -41,12 +41,6 @@
             <version>${version}</version>
         </dependency>
 
-        <dependency>
-            <groupId>com.alibaba</groupId>
-            <artifactId>rsqldb-common</artifactId>
-            <version>${version}</version>
-        </dependency>
-
 
         <dependency>
             <groupId>org.assertj</groupId>
